+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Changed logging configuration - log all classes at the level specified while building
 - Consent uses the loading of HTML classes from the file for some elements
 - Updated templates for consent page
+- Moved logs of user logins into the statistics filter, removed userID from it
 ### Fixed
 - Fixed handling of group "members" in resource capabilities
 - Fixed wrong behavior of MFA forced logout